(squash) people: Type user-related functions.
This add type annotations that are more or less related to a person, or
a "User". The types for User, CrossRealmBot, PeopleParams, ProfileData
are all created according to the OpenAPI Spec for the corresponding
schemas.

This also types the global dicts that we use throughout. Functions that
involve messages have their own complexity and are left out from this
batch of changes.

`_add_user` currently can also be called with `extract_people_from_message`
that adds a user (for deactivated users that aren't originally populated)
containing only a subset of required fields on `User`. The involvement of
these objects makes it type-unsafe whenever the global mappings for user
storage are accessed.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>